+++ domain/src/main/java/com/onlydust/marketplace/indexer/domain/services/exposers/PullRequestContributionExposer.java
@@ -1,4 +1,4 @@
-package com.onlydust.marketplace.indexer.domain.services;
+package com.onlydust.marketplace.indexer.domain.services.exposers;
 
 import com.onlydust.marketplace.indexer.domain.models.clean.CleanPullRequest;
 import com.onlydust.marketplace.indexer.domain.models.exposition.Contribution;
